FIXED: the idt desc propertie's byte sizes

+++ src/idt/idt.h
@@ -6,15 +6,15 @@ struct idt_desc
 {
     uint16_t offset_1; // Offset bits 0 - 15
     uint16_t selector;
-    uint16_t zero; //Unused set to zero
-    uint16_t type_attr;
+    uint8_t zero; //Unused set to zero
+    uint8_t type_attr;
     uint16_t offset_2; // Offset buts 16 - 31
 } __attribute__((packed));
 
 struct idtr_desc
 {
     uint16_t limit; // size of the descriptor table - 1
-    uint16_t base; // base address of the start of the IDT
+    uint32_t base; // base address of the start of the IDT
 } __attribute__((packed));
 
 void idt_init();
